requirements.txt
asgiref==3.2.5
astroid==2.4.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coverage==5.0.4
coveralls==1.11.1
cffi==1.14.3
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-cleanup==4.0.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-fsm==2.7.0
django-heroku==0.3.1
django-ranged-fileresponse==0.1.2
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
django-rest-knox==4.1.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
docopt==0.6.2
gunicorn==20.0.4
humanize==2.0.0
idna==2.9
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pymongo==3.11.0
pycparser==2.20
pyjwt==1.7.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.12.1
